Separated the learning variables to Theta1 and Theta2, Delta1 and Delta2, and so on. Implemented the algorithm again. Corrected a bag in adding avg quote price for the last date. Changed EPSILON fron 1/10^-4 to 0.12. Currently getting Cost in CrossValidat set = 1.4492997532205776.
Added three methods in Matrix operations for appending matrices and checking for NaN.
Fixed a bag in Training_X_Y_matrix - the case when we skipped all dates of a company as irrelevant. Also changed the Y matrix to consist of array of boolean pairs (0,1 if the price goes up, 1,0 otherwise).